requirements.txt
fastapi==0.110.0
uvicorn==0.15.0
requests==2.22.0
beautifulsoup4==4.7.1
lxml==4.7.1
pydantic==2.6.3
aiohttp==3.8.1
gql==3.0.0
feedparser==6.0.11
psycopg2-binary==2.9.9
regex==2023.10.3
numpy==1.23.5
meilisearch==0.31.0
requests_toolbelt==0.10.1
google-cloud-storage==1.33.0
fastapi-cache2==0.2.1
redis==4.6.0
pytz==2023.3.post1
meilisearch